+++ src/app/view-config-upgraders.js
@@ -389,3 +389,15 @@ export function upgradeFrom1_0_10(config) {
     version: '1.0.11',
   };
 }
+
+// Added in version 1.0.12:
+// - Added a fileType-to-dataType mapping
+// so that datasets[].files[].type is no longer required.
+export function upgradeFrom1_0_11(config) {
+  const newConfig = cloneDeep(config);
+
+  return {
+    ...newConfig,
+    version: '1.0.12',
+  };
+}