+++ src/loaders/types.js
@@ -1,5 +1,7 @@
 import { FileType } from '../app/constants';
-import { getPluginFileType, getPluginFileTypes } from '../app/plugins';
+import {
+  getLoaderClassesForPluginFileType,
+} from '../app/plugins';
 
 import JsonLoader from './JsonLoader';
 import MatrixZarrLoader from './MatrixZarrLoader';
@@ -32,17 +34,10 @@ export function getSourceAndLoaderFromFileType(type) {
   if (fileTypeToLoaderAndSource[type]) {
     return fileTypeToLoaderAndSource[type];
   }
-  const pluginFileType = getPluginFileType(type);
+  const pluginFileType = getLoaderClassesForPluginFileType(type);
   if (pluginFileType) {
     return pluginFileType;
   }
   // Fallback to JSON.
   return [JsonSource, JsonLoader];
-}
-
-export function getFileTypes() {
-  return [
-    ...Object.keys(fileTypeToLoaderAndSource),
-    ...getPluginFileTypes(),
-  ];
 }